Pin simpleeval to restore Python 2.6 compatibility*
While working on an unreleased Python project that uses rotate-backups
I noticed that the tox build for Python 2.6 was broken. Whether it's
worth it for me to keep supporting Python 2.6 is a valid question,
but right now the readme and setup script imply compatibility with
Python 2.6 so I feel half obliged to `fix this issue' :-).