Add an option `--use-rmdir` to use `rmdir` instead of `rm -R`
This is useful when the backups are CephFS snapshots.
Those are usually located in a special hidden directory (normally `.snap`) and cannot be removed by rm -R
(because they're read-only) but they can be removed by calling rmdir directly on the non-empty directory.
For example a simple backup procedure could be :
* mount cephfs to /backup
* backup the system to /backup
* make a snapshot with `mkdir /backup/.snap/<date>`
* rotate snapshots with `rotate-backups --use-rmdir /backup/.snap/`