Add UNIX timestamp extraction
Some backup solutions prefer using UNIX timestamps for tagging backup
files or directories. In order to enable the rotation of such
files/directories an extra switch is added to enable scanning for UNIX
timestamps in file- or directory-names.

If a number is found and a ValueError is raised, the extractor tries if
it is a valid milliseconds-timestamp.